# embeddings: add scoring function