--- a/enterprise/internal/embeddings/similarity_search.go
+++ b/enterprise/internal/embeddings/similarity_search.go
@@ -2,6 +2,7 @@ package embeddings
 
 import (
 	"container/heap"
+	"fmt"
 	"math"
 	"sort"
 
@@ -9,8 +10,8 @@ import (
 )
 
 type nearestNeighbor struct {
-	index      int
-	similarity float32
+	index int
+	score score
 }
 
 type nearestNeighborsHeap struct {
@@ -20,7 +21,7 @@ type nearestNeighborsHeap struct {
 func (nn *nearestNeighborsHeap) Len() int { return len(nn.neighbors) }
 
 func (nn *nearestNeighborsHeap) Less(i, j int) bool {
-	return nn.neighbors[i].similarity < nn.neighbors[j].similarity
+	return nn.neighbors[i].score.score < nn.neighbors[j].score.score
 }
 
 func (nn *nearestNeighborsHeap) Swap(i, j int) {
@@ -81,13 +82,15 @@ type WorkerOptions struct {
 
 // SimilaritySearch finds the `nResults` most similar rows to a query vector. It uses the cosine similarity metric.
 // IMPORTANT: The vectors in the embedding index have to be normalized for similarity search to work correctly.
-func (index *EmbeddingIndex[T]) SimilaritySearch(query []float32, numResults int, workerOptions WorkerOptions) []*T {
+func (index *EmbeddingIndex[T]) SimilaritySearch(query []float32, numResults int, workerOptions WorkerOptions, debug bool) SimilaritySearchResult[T] {
 	if numResults == 0 {
-		return []*T{}
+		return SimilaritySearchResult[T]{
+			RowMetadata: []*T{},
+		}
 	}
 
 	numRows := len(index.RowMetadata)
-	// Cannot request more results then there are rows.
+	// Cannot request more results than there are rows.
 	numResults = min(numRows, numResults)
 	// We need at least 1 worker.
 	numWorkers := max(1, workerOptions.NumWorkers)
@@ -101,12 +104,14 @@ func (index *EmbeddingIndex[T]) SimilaritySearch(query []float32, numResults int
 		for workerIdx := 0; workerIdx < len(rowsPerWorker); workerIdx++ {
 			// Capture the loop variable value so we can use it in the closure below.
 			workerIdx := workerIdx
-			wg.Go(func() { heaps[workerIdx] = index.partialSimilaritySearch(query, numResults, rowsPerWorker[workerIdx]) })
+			wg.Go(func() {
+				heaps[workerIdx] = index.partialSimilaritySearch(query, numResults, rowsPerWorker[workerIdx], debug)
+			})
 		}
 		wg.Wait()
 	} else {
 		// Run the similarity search directly when we have a single worker to eliminate the concurrency overhead.
-		heaps[0] = index.partialSimilaritySearch(query, numResults, rowsPerWorker[0])
+		heaps[0] = index.partialSimilaritySearch(query, numResults, rowsPerWorker[0], debug)
 	}
 
 	// Collect all heap neighbors from workers into a single array.
@@ -116,18 +121,27 @@ func (index *EmbeddingIndex[T]) SimilaritySearch(query []float32, numResults int
 			neighbors = append(neighbors, heap.neighbors...)
 		}
 	}
-	// And re-sort it according to the similarity (descending).
-	sort.Slice(neighbors, func(i, j int) bool { return neighbors[i].similarity > neighbors[j].similarity })
+	// And re-sort it according to the score (descending).
+	sort.Slice(neighbors, func(i, j int) bool { return neighbors[i].score.score > neighbors[j].score.score })
 
 	// Take top neighbors and return them as results.
-	results := make([]*T, numResults)
+	results := SimilaritySearchResult[T]{
+		RowMetadata: make([]*T, numResults),
+	}
+
+	if debug {
+		results.Debug = make([]string, numResults)
+	}
 	for idx := 0; idx < min(numResults, len(neighbors)); idx++ {
-		results[idx] = &index.RowMetadata[neighbors[idx].index]
+		results.RowMetadata[idx] = &index.RowMetadata[neighbors[idx].index]
+		if debug {
+			results.Debug[idx] = neighbors[idx].score.debug
+		}
 	}
 	return results
 }
 
-func (index *EmbeddingIndex[T]) partialSimilaritySearch(query []float32, numResults int, partialRows partialRows) *nearestNeighborsHeap {
+func (index *EmbeddingIndex[T]) partialSimilaritySearch(query []float32, numResults int, partialRows partialRows, debug bool) *nearestNeighborsHeap {
 	nRows := partialRows.end - partialRows.start
 	if nRows <= 0 {
 		return nil
@@ -136,27 +150,66 @@ func (index *EmbeddingIndex[T]) partialSimilaritySearch(query []float32, numResu
 
 	nnHeap := newNearestNeighborsHeap()
 	for i := partialRows.start; i < partialRows.start+numResults; i++ {
-		similarity := CosineSimilarity(
-			index.Embeddings[i*index.ColumnDimension:(i+1)*index.ColumnDimension],
-			query,
-		)
-		heap.Push(nnHeap, nearestNeighbor{i, similarity})
+		heap.Push(nnHeap, nearestNeighbor{i, index.score(query, i, debug)})
 	}
 
 	for i := partialRows.start + numResults; i < partialRows.end; i++ {
-		similarity := CosineSimilarity(
-			index.Embeddings[i*index.ColumnDimension:(i+1)*index.ColumnDimension],
-			query,
-		)
+		score := index.score(query, i, debug)
 		// Add row if it has greater similarity than the smallest similarity in the heap.
 		// This way we ensure keep a set of highest similarities in the heap.
-		if similarity > nnHeap.Peek().similarity {
+		if score.score > nnHeap.Peek().score.score {
 			heap.Pop(nnHeap)
-			heap.Push(nnHeap, nearestNeighbor{i, similarity})
+			heap.Push(nnHeap, nearestNeighbor{i, score})
 		}
 	}
 
 	return nnHeap
+}
+
+type score struct {
+	score float32
+	debug string
+}
+
+const (
+	scoreFileRankWeight   float32 = 0.5
+	scoreSimilarityWeight float32 = 0.5
+)
+
+func (index *EmbeddingIndex[T]) score(query []float32, i int, debug bool) score {
+	score := score{}
+	addScore := func(what string, s float32) {
+		score.score += s
+		if debug {
+			score.debug += fmt.Sprintf("%s:%.2f, ", what, s)
+		}
+	}
+
+	similarity := CosineSimilarity(
+		index.Embeddings[i*index.ColumnDimension:(i+1)*index.ColumnDimension],
+		query,
+	)
+
+	addScore("similarity", scoreSimilarityWeight*similarity)
+
+	// handle missing ranks
+	if len(index.Ranks) > i {
+		// The file rank represents a log (base 2) count. The log ranks should be
+		// bounded at 32, but we cap it just in case to ensure it falls in the range [0,
+		// 1]. I am not using math.Min here to avoid the back and forth conversion
+		// between float64 adn float32.
+		normalizedRank := index.Ranks[i] / 32.0
+		if normalizedRank > 1.0 {
+			normalizedRank = 1.0
+		}
+		addScore("rank", scoreFileRankWeight*normalizedRank)
+	}
+
+	if debug {
+		score.debug = fmt.Sprintf("score: %.2f, %s", score.score, score.debug)
+	}
+
+	return score
 }
 
 func CosineSimilarity(row []float32, query []float32) float32 {