--- a/enterprise/internal/embeddings/types.go
+++ b/enterprise/internal/embeddings/types.go
@@ -4,29 +4,24 @@ import (
 	"github.com/sourcegraph/sourcegraph/internal/api"
 )
 
-type EmbeddingIndex[T any] struct {
+type EmbeddingIndex struct {
 	Embeddings      []float32
 	ColumnDimension int
-	RowMetadata     []T
+	RowMetadata     []RepoEmbeddingRowMetadata
 	Ranks           []float32
 }
 
-type SimilaritySearchResult[T any] struct {
-	RowMetadata []*T
-	Debug       []string
-}
-
 type RepoEmbeddingRowMetadata struct {
-	FileName  string
-	StartLine int
-	EndLine   int
+	FileName  string `json:"fileName"`
+	StartLine int    `json:"startLine"`
+	EndLine   int    `json:"endLine"`
 }
 
 type RepoEmbeddingIndex struct {
 	RepoName  api.RepoName
 	Revision  api.CommitID
-	CodeIndex EmbeddingIndex[RepoEmbeddingRowMetadata]
-	TextIndex EmbeddingIndex[RepoEmbeddingRowMetadata]
+	CodeIndex EmbeddingIndex
+	TextIndex EmbeddingIndex
 }
 
 type ContextDetectionEmbeddingIndex struct {
@@ -40,10 +35,8 @@ type EmbeddingSearchResults struct {
 }
 
 type EmbeddingSearchResult struct {
-	FileName  string `json:"fileName"`
-	StartLine int    `json:"startLine"`
-	EndLine   int    `json:"endLine"`
-	Content   string `json:"content"`
+	RepoEmbeddingRowMetadata
+	Content string `json:"content"`
 	// Experimental: Clients should not rely on any particular format of debug
 	Debug string `json:"debug,omitempty"`
 }