related to #49710

This adds a scoring function which computes the weighted sum of cosine similarity and document rank.

For now, the similarity and file rank weights are in a  2:1 ratio. 

Anecdotically, I saw in my experiments how a 'go.sum' file (with high similarity) was ranked down because of its low document rank.

I also added a new field "debug" (bool) to the API which, similarily to Zoekt, adds debug information about the score to the response if set to `true`.

The API change is backward compatible.

## Test plan
- new unit test
- manual testing:
  - Tested embeddings API for the following scenarios
    - repo has ranks but no embeddings -> error
    - repo has embeddings but no document ranks -> reponse with ranking based soley on embeddings
    - repo has embeddings and document ranks   -> response with ranking based on weighted sum
  - Confirmed that not setting `Debug=true` omits `Debug` from the response
  - Confirmed that "old" indexes continue to work