@stefanhengl 2023-03-28
Compare benchmarks main vs. this branch => no statistically significant difference.
```
goos: darwin
goarch: arm64
pkg: github.com/sourcegraph/sourcegraph/enterprise/internal/embeddings
                                  │  main.txt   │      add-scoring-function.txt      │
                                  │   sec/op    │   sec/op     vs base               │
EmbeddingIndexStorage-10            558.4m ± 1%   562.2m ± 1%       ~ (p=0.436 n=10)
SimilaritySearch/numWorkers=1-10     1.872 ± 1%    1.879 ± 1%       ~ (p=0.075 n=10)
SimilaritySearch/numWorkers=2-10    948.3m ± 2%   946.9m ± 2%       ~ (p=0.529 n=10)
SimilaritySearch/numWorkers=4-10    493.6m ± 0%   494.4m ± 0%       ~ (p=0.190 n=10)
SimilaritySearch/numWorkers=8-10    276.0m ± 7%   273.0m ± 8%       ~ (p=0.684 n=10)
SimilaritySearch/numWorkers=16-10   268.7m ± 4%   272.7m ± 7%       ~ (p=0.481 n=10)
geomean                             575.4m        576.8m       +0.25%
```